web/backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
sqlalchemy>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
cryptography>=41.0.0
python-dotenv>=0.20.0
aiosqlite>=0.19.0
httpx>=0.24.0
requests>=2.28.0
pandas>=1.5.0
numpy>=1.23.0
pytz>=2022.1
